Polymorphism
Polymorphism means many forms, we use polymorphism to write generic code based on the base class or a parent class. And this code in the main method is extendable, meaning, it doesn't have to change as new subclasses become available. This code can handle any instances that are a move or subclass of movie that are returned for the factory method.